separated views into separate files
