+++ packages/views/order_edit.py
@@ -1,124 +1,8 @@
 from django.core import serializers
-from django.shortcuts import render, redirect
-from .models import Order, City, Address, Street
-from .models import Partner
-from .models import Package
+from packages.models import Address, City, Order, Package, Partner, Street
+from django.shortcuts import render
 from util import is_member, is_staff
-
-
-def partner_search(request, partner_name):
-    partner = Partner.objects.filter(name=partner_name).get()
-
-    name_query = request.GET.get('name')
-    street_name_query = request.GET.get('street_name')
-    street_number_query = request.GET.get('street_number')
-    city_query = request.GET.get('city')
-    number_query = request.GET.get('number')
-
-    packages = Package.objects.filter(order__partner__name=partner.name)
-    filtered_packages = Package.objects.none()
-
-    if name_query:
-        packages = packages.filter(full_name__icontains=name_query)
-        filtered_packages = packages
-    if number_query:
-        packages = packages.filter(phone_number__icontains=number_query)
-        filtered_packages = packages
-    if street_name_query:
-        packages = packages.filter(destination__street__name__icontains=street_name_query)
-        filtered_packages = packages
-    if street_number_query:
-        packages = packages.filter(destination__street_number__contains=street_number_query)
-        filtered_packages = packages
-    if city_query:
-        packages = packages.filter(destination__city__name__icontains=city_query)
-        filtered_packages = packages
-
-    orders = []
-    related_packages = []
-    for package in filtered_packages.order_by('-order__collection_date'):
-        order = package.order
-        if order not in orders:
-            orders.append(order)
-            related_packages.append([package])
-        else:
-            related_packages[orders.index(order)].append(package)
-
-    order_packages = [(orders[i], related_packages[i]) for i in range(0, len(orders))]
-    context = {
-        'partner': partner,
-        'order_packages': order_packages,
-        'is_staff': is_staff(request.user)
-    }
-    context.update(string_data_lists_context())
-
-    return render(request, 'packages/partner_search.html', context)
-
-
-def order_view(request, order_id):
-    order = Order.objects.get(id=order_id)
-
-    if request.POST:
-        if request.POST.get('package') == 'all':
-            for inner in order.related_packages():
-                inner.as_query().update(status=inner.next_status())
-        else:
-            package_id = request.POST.get('package')
-            package = order.related_packages().filter(id=package_id)
-            package.update(status=package.get().next_status())
-
-    packages = list(order.related_packages())
-    context = {
-        'order': order,
-        'packages': packages,
-        'is_staff': is_staff(request.user)
-    }
-
-    return render(request, "packages/order.html", context)
-
-
-def package_view(request, package_id):
-    package = Package.objects.get(id=package_id)
-    name = package.full_name.split(' ')[0] + ':'
-
-    if not name:
-        name = ''
-
-    context = {
-        'package': package,
-        'name': name,
-        'is_staff': is_staff(request.user)
-    }
-
-    return render(request, "packages/package.html", context)
-
-
-def partner_view(request, partner_name):
-    requested_partner = Partner.objects.get(name=partner_name)
-
-    if not is_member(request.user, requested_partner.name) and not is_staff(request.user):
-        return render(request, 'errors/access_restricted.html', {})
-
-    if request.POST:
-        new_order_id = str(Order.objects.create(partner=requested_partner).id)
-        return redirect('/partners/'+requested_partner.name+'/'+new_order_id+'/')
-    orders = requested_partner.related_orders().order_by('-collection_date')
-    package_amounts = []
-    order_statuses = []
-
-    for order in orders:
-        package_amounts.append(len(order.related_packages()))
-        order_statuses.append(order.overall_package_status())
-
-    order_amount_status = [(orders[i], package_amounts[i], order_statuses[i]) for i in range(0, len(orders))]
-    context = {
-        'order_amount_status': order_amount_status,
-        'is_staff': is_staff(request.user),
-        'partner': requested_partner
-    }
-
-    return render(request, "packages/partner.html", context)
-
+from packages.util import string_data_lists_context
 
 def order_edit_view(request, partner_name, order_id):
     partner = Partner.objects.get(name=partner_name)
@@ -159,18 +43,6 @@ def order_edit_view(request, partner_name, order_id):
     return response
 
 
-def staff_view(request):
-    if not is_member(request.user, 'staff'):
-        return render(request, 'errors/access_restricted.html', {})
-
-    context = {
-        'partners': Partner.objects.all(),
-        'is_staff': True
-    }
-
-    return render(request, 'packages/staff.html', context)
-
-
 def change_packages_status(package_id, update_type, order):
     package = Package.objects.filter(id=package_id)
 
@@ -184,7 +56,6 @@ def change_packages_status(package_id, update_type, order):
     else:
         for inner in order.related_packages():
             inner.as_query().update(status=inner.prev_status())
-
 
 def update_saved_packages(request):
     packages = serializers.deserialize('json', request.COOKIES.get('updated_packages'))
@@ -267,14 +138,6 @@ def get_or_create_city(name, area):
     if not city.exists():
         return City.objects.create(name=name, area=area)
     return city.get()
-
-
-def string_data_lists_context():
-    return {
-        'cities': City.objects.all(),
-        'streets': Street.objects.all(),
-        'names': Package.objects.exclude(full_name='').values_list('full_name', flat=True)
-    }
 
 
 def add_package_to_json(package, json):
